chore: move class and decoator to one file
